requirements.txt
email-validator==1.2.1
Flask==2.2.2
Flask-WTF==1.0.1
passlib==1.7.4
python-dotenv==0.20.0
WTForms==3.0.1
gunicorn==20.1.0
pymongo==4.2.0
itsdangerous==2.1.2
